requirements.txt
albumentations==1.2.1
autograd==1.5
autopep8==2.0.0
glob2==0.7
imageio==2.9.0
matplotlib==3.2.2
numpy==1.21.6
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
pandas==1.3.5
pandas-datareader==0.9.0
pandas-gbq==0.17.9
pandas-profiling==1.4.1
pathlib==1.0.1
Pillow==7.1.2
pip-tools==6.2.0
py==1.11.0
python-apt==0.0.0
python-dateutil==2.8.2
python-louvain==0.16
python-slugify==7.0.0
python-utils==3.4.5
PyYAML==6.0
pyzmq==23.2.1
scikit-image==0.18.3
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
sklearn-pandas==1.8.0
timm==0.6.12
torch @ https://download.pytorch.org/whl/cu116/torch-1.13.0%2Bcu116-cp38-cp38-linux_x86_64.whl
torchaudio @ https://download.pytorch.org/whl/cu116/torchaudio-0.13.0%2Bcu116-cp38-cp38-linux_x86_64.whl
torchsummary==1.5.1
torchtext==0.14.0
torchvision @ https://download.pytorch.org/whl/cu116/torchvision-0.14.0%2Bcu116-cp38-cp38-linux_x86_64.whl
tqdm==4.64.1
xgboost==0.90